requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
bleach==6.0.0
    # via galaxy-util
boltons==21.0.0
    # via galaxy-util
certifi==2022.12.7
    # via requests
charset-normalizer==3.0.1
    # via requests
docutils==0.19
    # via galaxy-util
galaxy-containers==22.1.1
    # via galaxy-tool-util
galaxy-tool-util==22.1.5
    # via -r requirements.in
galaxy-util==22.1.2
    # via
    #   galaxy-containers
    #   galaxy-tool-util
idna==3.4
    # via requests
importlib-resources==5.10.2
    # via galaxy-util
lxml==4.9.2
    # via galaxy-tool-util
markupsafe==2.1.2
    # via galaxy-util
packaging==21.3
    # via
    #   -r requirements.in
    #   galaxy-util
pycryptodome==3.17
    # via galaxy-util
pydantic==1.10.4
    # via galaxy-tool-util
pyparsing==3.0.9
    # via
    #   galaxy-util
    #   packaging
pyyaml==6.0
    # via
    #   galaxy-tool-util
    #   galaxy-util
repoze-lru==0.7
    # via routes
requests==2.28.2
    # via
    #   galaxy-containers
    #   galaxy-util
routes==2.5.1
    # via galaxy-util
six==1.16.0
    # via
    #   bleach
    #   routes
sortedcontainers==2.4.0
    # via galaxy-tool-util
typing-extensions==4.4.0
    # via
    #   galaxy-tool-util
    #   galaxy-util
    #   pydantic
urllib3==1.26.14
    # via requests
webencodings==0.5.1
    # via bleach
zipstream-new==1.1.8
    # via galaxy-util
